doctor & request authorization

+++ flaskr/services/request_service.py
@@ -1,4 +1,4 @@
-from flaskr.models import PatientRequest
+from flaskr.models import Patient, PatientRequest
 from flaskr.extensions import db
 
 def add_patient_request(patient_id, doctor_id):
@@ -6,15 +6,26 @@ def add_patient_request(patient_id, doctor_id):
         .filter_by(patient_id=patient_id, doctor_id=doctor_id)
         .first()):
         raise ValueError(f'patient with ID: {patient_id} already requested doctor with ID: {doctor_id}')
+    if (Patient.query
+        .filter_by(user_id=patient_id)
+        .first().doctor_id == doctor_id):
+        raise ValueError(f'patient with ID: {patient_id} is already assigned doctor with ID: {doctor_id}')
     new_request = PatientRequest(patient_id=patient_id, doctor_id=doctor_id)
     db.session.add(new_request)
     db.session.commit()
     return new_request.to_dict()
 
-def delete_patient_request(request_id):
+def delete_patient_request(request_id, requesting_user=None):
+    from flaskr.services import UnauthorizedError
+    if not requesting_user:
+        raise UnauthorizedError
     request = PatientRequest.query.filter_by(request_id=request_id).first()
     if not request:
         return None
+    if not (requesting_user.account_type.name != 'SuperUser' 
+            and request.doctor_id == requesting_user.user_id):
+        raise UnauthorizedError
+
     db.session.delete(request)
     db.session.commit()
     return request.to_dict()
